# Conflicts:
#	.idea/dataSources.xml
#	flowback/poll/models.py
#	flowback/poll/tests/test_poll.py
#	flowback/poll/views/poll.py

+++ flowback/group/services.py
@@ -26,7 +26,7 @@ group_schedule = ScheduleManager(schedule_origin_name='group', possible_origins=
 group_kanban = KanbanManager(origin_type='group')
 group_notification = NotificationManager(sender_type='group', possible_categories=['group', 'members', 'invite',
                                                                                    'delegate', 'poll', 'kanban',
-                                                                                   'schedule'])
+                                                                                   'schedule', 'poll_schedule'])
 
 
 def group_notification_subscribe(*, user_id: int, group: int, categories: list[str]):
@@ -99,7 +99,8 @@ def group_permission_update(*, user: int, group: int, permission_id: int, data) 
     non_side_effect_fields = ['role_name',
                               'invite_user',
                               'create_poll',
-                              'poll_quorum'
+                              'poll_fast_forward',
+                              'poll_quorum',
                               'allow_vote',
                               'kick_members',
                               'ban_members',
@@ -198,7 +199,7 @@ def group_leave(*, user: int, group: int) -> None:
 
 def group_invite(*, user: int, group: int, to: int) -> GroupUserInvite:
     group = group_user_permissions(group=group, user=user, permissions=['admin', 'invite_user']).group
-    get_object(GroupUser, error_message='User is already in the group', reverse=True, group=group, user=to)
+    get_object(GroupUser, error_message='User is already in the group', reverse=True, group=group, user=to, active=True)
     invite = GroupUserInvite(user_id=to, group_id=group.id, external=False)
 
     invite.full_clean()
@@ -216,14 +217,22 @@ def group_invite_remove(*, user: int, group: int, to: int) -> None:
 def group_invite_accept(*, fetched_by: int, group: int, to: int = None) -> None:
     if to:
         group_user_permissions(group=group, user=fetched_by, permissions=['invite_user', 'admin'])
-        get_object(GroupUser, 'User already joined', reverse=True, user_id=to, group=group)
+        get_object(GroupUser, 'User already joined', reverse=True, user_id=to, group=group, active=True)
         invite = get_object(GroupUserInvite, 'User has not requested invite', user_id=to, group_id=group, external=True)
-        group_user = GroupUser(user_id=to, group_id=group)
+
+        if group_user := get_object(GroupUser, raise_exception=False, user_id=to, group_id=group, active=False):
+            group_user.active = True
+        else:
+            group_user = GroupUser(user_id=to, group_id=group)
 
     else:
         invite = get_object(GroupUserInvite, 'User has not been invited', user_id=fetched_by, group_id=group,
                             external=False)
-        group_user = GroupUser(user_id=fetched_by, group_id=group)
+
+        if group_user := get_object(GroupUser, raise_exception=False, user_id=fetched_by, group_id=group, active=False):
+            group_user.active = True
+        else:
+            group_user = GroupUser(user_id=fetched_by, group_id=group)
 
     # TODO fix group_user.full_clean()
     group_user.save()
@@ -237,7 +246,12 @@ def group_invite_reject(*, fetched_by: id, group: int, to: int = None) -> None:
         invite = get_object(GroupUserInvite, 'User has not been invited', user_id=to, group_id=group)
 
     else:
-        get_object(GroupUser, 'User already joined', reverse=True, user_id=fetched_by, group_id=group)
+        get_object(GroupUser,
+                   'User already joined',
+                   reverse=True,
+                   user_id=fetched_by,
+                   group_id=group,
+                   active=True)
         invite = get_object(GroupUserInvite, 'User has not requested invite', user_id=fetched_by, group_id=group)
 
     invite.delete()
@@ -257,6 +271,11 @@ def group_tag_update(*, user: int, group: int, tag: int, data) -> GroupTags:
     tag = get_object(GroupTags, group_id=group, id=tag)
     non_side_effect_fields = ['active']
 
+    if (GroupTags.objects.filter(group_id=group, active=True).count() <= 1
+            and tag.is_active
+            and data.get('active')):
+        raise ValidationError('Group must have at least one active tag available for users')
+
     group_tag, has_updated = model_update(instance=tag,
                                           fields=non_side_effect_fields,
                                           data=data)
@@ -267,6 +286,10 @@ def group_tag_update(*, user: int, group: int, tag: int, data) -> GroupTags:
 def group_tag_delete(*, user: int, group: int, tag: int) -> None:
     group_user_permissions(group=group, user=user, permissions=['admin'])
     tag = get_object(GroupTags, group_id=group, id=tag)
+
+    if GroupTags.objects.filter(group_id=group, active=True).count() <= 1 and tag.is_active:
+        raise ValidationError('Group must have at least one active tag available for users')
+
     tag.delete()
 
 
@@ -417,7 +440,7 @@ def group_kanban_entry_create(*,
                               fetched_by_id: int,
                               assignee_id: int = None,
                               title: str,
-                              description: str,
+                              description: str = None,
                               priority: int,
                               tag: int,
                               end_date: timezone.datetime = None
@@ -537,3 +560,53 @@ def group_thread_comment_delete(user_id: int, thread_id: int, comment_id: int):
     return comment_delete(fetched_by=user_id,
                           comment_section_id=thread.comment_section_id,
                           comment_id=comment_id)
+
+
+def group_delegate_pool_comment_create(*,
+                                       author_id: int,
+                                       delegate_pool_id: int,
+                                       message: str,
+                                       attachments: list = None,
+                                       parent_id: int = None) -> Comment:
+    delegate_pool = get_object(GroupUserDelegatePool, id=delegate_pool_id)
+    group_user_permissions(group=delegate_pool.group, user=author_id)
+
+    comment = comment_create(author_id=author_id,
+                             comment_section_id=delegate_pool.comment_section.id,
+                             message=message,
+                             parent_id=parent_id,
+                             attachments=attachments,
+                             attachment_upload_to="group/delegate_pool/comment/attachments")
+
+    return comment
+
+
+def group_delegate_pool_comment_update(*,
+                                       author_id: int,
+                                       delegate_pool_id: int,
+                                       comment_id: int,
+                                       data) -> Comment:
+    delegate_pool = get_object(GroupUserDelegatePool, id=delegate_pool_id)
+    group_user_permissions(group=delegate_pool.group, user=author_id)
+
+    return comment_update(fetched_by=author_id,
+                          comment_section_id=delegate_pool.comment_section.id,
+                          comment_id=comment_id,
+                          data=data)
+
+
+def group_delegate_pool_comment_delete(*,
+                                       author_id: int,
+                                       delegate_pool_id: int,
+                                       comment_id: int):
+    delegate_pool = get_object(GroupUserDelegatePool, id=delegate_pool_id)
+    group_user = group_user_permissions(group=delegate_pool.group, user=author_id)
+
+    force = bool(group_user_permissions(group_user=group_user,
+                                        permissions=['admin', 'force_delete_comment'],
+                                        raise_exception=False))
+
+    return comment_delete(fetched_by=author_id,
+                          comment_section_id=delegate_pool.comment_section.id,
+                          comment_id=comment_id,
+                          force=force)
